backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
PyJWT==2.8.0
pymysql==1.1.0
cryptography==42.0.5
python-dotenv==1.0.0
bcrypt==4.0.1
werkzeug==2.3.8 